Add a research tree load type option: Do not generate research tree, but keep research queue
This option is displayed only if the OverrideResearch checkbox is disabled.
It disables the initialization of the research tree and the display of the research tree window, leaving only the research queue functionality. This allows you to queue research in the vanilla research window by Ctrl/Shift + Left/Right Click.